Remove stevedore package requirement
Avocado uses the stevedore library, which has origins in OpenStack, as
a way to find plugins and activate them.  It's very useful, but
Avocado itself only uses a very small part of it, and actually extends
most of what it uses.

Given that stevedore itself is a simple wrapper around setuptools
(pkg_resources, really), this simply adds the little code we need
around pkg_resources.

The benefits is that the core Avocado application now depends only on
Python and setuptools (which is bundled together on many systems).

Reference: https://trello.com/c/vCgzJ02N
