+++ python-avocado.spec
@@ -90,7 +90,6 @@ BuildRequires: python3-lxml
 BuildRequires: python3-psutil
 BuildRequires: python3-resultsdb_api
 BuildRequires: python3-setuptools
-BuildRequires: python3-stevedore
 BuildRequires: python3-pycdlib
 
 %if %{with_tests}
@@ -116,7 +115,6 @@ Requires: gdb-gdbserver
 Requires: procps-ng
 Requires: python3
 Requires: python3-setuptools
-Requires: python3-stevedore
 Requires: python3-pycdlib
 
 %description -n python3-%{srcname}
